requirements.txt
opencv-python-headless
pillow
numpy
streamlit
